+++ test/commands/multi.spec.js
@@ -65,6 +65,31 @@ describe("The 'multi' method", function () {
                     multi1.get('m1');
                     multi1.exec(done);
                 });
+
+                it("executes a pipelined multi properly after a reconnect in combination with the offline queue", function (done) {
+                    client.once('ready', function () {
+                        client.stream.destroy();
+                        var called = false;
+                        var multi1 = client.multi();
+                        multi1.set("m1", "123");
+                        multi1.get('m1');
+                        multi1.exec(function (err, res) {
+                            assert(!err);
+                            called = true;
+                        });
+                        client.once('ready', function () {
+                            var multi1 = client.multi();
+                            multi1.set("m2", "456");
+                            multi1.get('m2');
+                            multi1.exec(function (err, res) {
+                                assert(called);
+                                assert(!err);
+                                assert.strictEqual(res[1], '456');
+                                done();
+                            });
+                        });
+                    });
+                });
             });
 
             describe("when connection is broken", function () {
@@ -535,6 +560,35 @@ describe("The 'multi' method", function () {
                         assert.strictEqual(input[1].length, 2);
                         done();
                     });
+                });
+
+                it("works properly after a reconnect. issue #897", function (done) {
+                    helper.serverVersionAtLeast.call(this, client, [2, 6, 5]);
+
+                    client.stream.destroy();
+                    client.on('error', function (err) {
+                        assert.strictEqual(err.code, 'ECONNREFUSED');
+                    });
+                    client.on('ready', function () {
+                        client.multi([['set', 'foo', 'bar'], ['get', 'foo']]).exec(function (err, res) {
+                            assert(!err);
+                            assert.strictEqual(res[1], 'bar');
+                            done();
+                        });
+                    });
+                });
+
+                it("emits error once if reconnecting after multi has been executed but not yet returned without callback", function (done) {
+                    helper.serverVersionAtLeast.call(this, client, [2, 6, 5]);
+
+                    client.on('error', function(err) {
+                        assert.strictEqual(err.code, 'UNCERTAIN_STATE');
+                        done();
+                    });
+
+                    client.multi().set("foo", 'bar').get('foo').exec();
+                    // Abort connection before the value returned
+                    client.stream.destroy();
                 });
 
             });